feat: Enhance Dashboard with Overview and Journal History components; add mock data for journaling insights
