+++ client/src/app/dashboard.tsx
@@ -3,16 +3,17 @@ import { Edit3, History } from 'lucide-react';
 import { useNavigate, useParams } from '@tanstack/react-router';
 import TodayJournal from '@/components/dashboard/TodayJournal';
 import JournalHistory from '@/components/dashboard/JournalHistory';
+import Overview from '@/components/dashboard/Overview';
 
 const Dashboard = () => {
   const params = useParams({ strict: false });
   const navigate = useNavigate();
 
   const activeTab = params.tab
-    ? (params.tab as 'today' | 'calendar' | 'history')
+    ? (params.tab as 'today' | 'overview' | 'history')
     : 'today';
 
-  const handleTabChange = (tab: 'today' | 'calendar' | 'history') => {
+  const handleTabChange = (tab: 'today' | 'overview' | 'history') => {
     if (tab === 'today') {
       navigate({ to: '/dashboard' });
     } else {
@@ -25,7 +26,7 @@ const Dashboard = () => {
       <div className="container mx-auto py-8">
         {/* Tab Navigation */}
         <div className="flex justify-center mb-8">
-          <div className="bg-white rounded-xl p-1 shadow-sm border border-gray-200 inline-flex">
+          <div className="bg-white rounded-xl p-1 shadow-sm border border-gray-200 inline-flex space-x-2">
             <motion.button
               onClick={() => handleTabChange('today')}
               className={`px-6 py-3 rounded-lg text-sm font-medium transition-all flex items-center space-x-2 ${
@@ -39,19 +40,6 @@ const Dashboard = () => {
               <Edit3 className="w-4 h-4" />
               <span>Today's Journal</span>
             </motion.button>
-            {/* <motion.button
-              onClick={() => setActiveTab('calendar')}
-              className={`px-6 py-3 rounded-lg text-sm font-medium transition-all flex items-center space-x-2 ${
-                activeTab === 'calendar'
-                  ? 'bg-teal-500 text-white shadow-md'
-                  : 'text-gray-600 hover:text-gray-900 hover:bg-gray-50'
-              }`}
-              whileHover={{ scale: 1.02 }}
-              whileTap={{ scale: 0.98 }}
-            >
-              <Calendar className="w-4 h-4" />
-              <span>Calendar</span>
-            </motion.button> */}
             <motion.button
               onClick={() => handleTabChange('history')}
               className={`px-6 py-3 rounded-lg text-sm font-medium transition-all flex items-center space-x-2 ${
@@ -65,21 +53,26 @@ const Dashboard = () => {
               <History className="w-4 h-4" />
               <span>Previous Journals</span>
             </motion.button>
+            <motion.button
+              onClick={() => handleTabChange('overview')}
+              className={`px-6 py-3 rounded-lg text-sm font-medium transition-all flex items-center space-x-2 ${
+                activeTab === 'overview'
+                  ? 'bg-teal-500 text-white shadow-md'
+                  : 'text-gray-600 hover:text-gray-900 hover:bg-gray-50'
+              }`}
+              whileHover={{ scale: 1.02 }}
+              whileTap={{ scale: 0.98 }}
+            >
+              <Edit3 className="w-4 h-4" />
+              <span>Overview</span>
+            </motion.button>
           </div>
         </div>
 
         <AnimatePresence mode="wait">
           {activeTab === 'today' ? (
             <TodayJournal />
-          ) : activeTab === 'calendar' ? (
-            <motion.div
-              key="calendar"
-              initial={{ opacity: 0, x: -20 }}
-              animate={{ opacity: 1, x: 0 }}
-              exit={{ opacity: 0, x: 20 }}
-              transition={{ duration: 0.3 }}
-            ></motion.div>
-          ) : (
+          ) : activeTab === 'history' ? (
             <motion.div
               key="history"
               initial={{ opacity: 0, x: -20 }}
@@ -88,6 +81,16 @@ const Dashboard = () => {
               transition={{ duration: 0.3 }}
             >
               <JournalHistory />
+            </motion.div>
+          ) : (
+            <motion.div
+              key="overview"
+              initial={{ opacity: 0, x: -20 }}
+              animate={{ opacity: 1, x: 0 }}
+              exit={{ opacity: 0, x: 20 }}
+              transition={{ duration: 0.3 }}
+            >
+              <Overview />
             </motion.div>
           )}
         </AnimatePresence>
